Se agrego un mensaje de confirmacion al eliminar un registro de las grillas

+++ src/pages/dashboard/pages_dashboard/cuits-restringidos/CuitsRestringidos.jsx
@@ -16,6 +16,7 @@ import DeleteIcon from "@mui/icons-material/DeleteOutlined";
 import SaveIcon from "@mui/icons-material/Save";
 import CancelIcon from "@mui/icons-material/Close";
 import { actualizarCuitRestringido, crearCuitRestringido, eliminarCuitRestringido, obtenerCuitsRestringidos } from './CuitsRestringidosApi';
+import Swal from 'sweetalert2';
 
 function EditToolbar(props) {
   const { setRows, rows, setRowModesModel } = props;
@@ -60,7 +61,7 @@ export const CuitsRestringidos = () => {
   );
 
   useEffect(() => {
-    
+
     const ObtenerCuitsRestringidos = async () => {
 
       const cuitsRestringidosResponse = await obtenerCuitsRestringidos(TOKEN);
@@ -90,9 +91,28 @@ export const CuitsRestringidos = () => {
 
   const handleDeleteClick = (id) => async () => {
 
-    setRows(rows.filter((row) => row.id !== id));
+    const showSwalConfirm = async () => {
+      try {
+        Swal.fire({
+          title: '¿Estás seguro?',
+          text: "¡No podrás revertir esto!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#1A76D2',
+          cancelButtonColor: '#6c757d',
+          confirmButtonText: 'Si, bórralo!'
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            setRows(rows.filter((row) => row.id !== id));
+            await eliminarCuitRestringido(id, TOKEN);
+          }
+        });
+      } catch (error) {
+        console.error('Error al ejecutar eliminarFeriado:', error);
+      }
+    };
 
-    await eliminarCuitRestringido(id, TOKEN);
+    showSwalConfirm();
   };
 
   const handleCancelClick = (id) => () => {
