+++ src/pages/dashboard/pages_dashboard/declaraciones_juradas/mis_declaraciones_juradas/grilla_mis_declaraciones_juradas/GrillaMisDeclaracionesJuradas.jsx
@@ -22,6 +22,7 @@ import {
 } from "./GrillaMisDeclaracionesJuradasApi";
 import { PDFViewer, PDFDownloadLink } from "@react-pdf/renderer";
 import { MyDocument } from "./MiPdf";
+import Swal from 'sweetalert2'
 
 /* function EditToolbar(props) {
 
@@ -117,9 +118,32 @@ export const GrillaMisDeclaracionesJuradas = ({
   };
 
   const handleDeleteClick = (id) => async () => {
-    setRowsMisDdjj(rows_mis_ddjj.filter((row) => row.id !== id));
+    
 
-    await eliminarDeclaracionJurada(idEmpresa, id, token);
+    const showSwalConfirm = async () => {
+      try {
+        Swal.fire({
+          title: '¿Estás seguro?',
+          text: "¡No podrás revertir esto!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#1A76D2',
+          cancelButtonColor: '#6c757d',
+          confirmButtonText: 'Si, bórralo!'
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+
+            setRowsMisDdjj(rows_mis_ddjj.filter((row) => row.id !== id));
+            
+            await eliminarDeclaracionJurada(idEmpresa, id, token);
+          }
+        });
+      } catch (error) {
+        console.error('Error al ejecutar eliminarFeriado:', error);
+      }
+    };
+
+    showSwalConfirm();
   };
 
   const handleCancelClick = (id) => () => {
